# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Hunting-Fishball Context Engine API
  description: |
    Public HTTP surface for the Hunting-Fishball multi-tenant
    knowledge / context platform. All `/v1/*` endpoints expect:

    - `X-Tenant-ID`: required, ULID-formatted tenant identifier
      (the auth middleware writes this into the gin context;
      handlers never read tenant identity from query / body).
    - `X-Actor-ID`: optional, ULID-formatted actor for audit logs.

    Liveness / readiness / metrics are served without auth on
    `/healthz`, `/readyz`, `/metrics`.
  version: 0.8.0
servers:
  - url: http://localhost:8080
    description: Local development
components:
  parameters:
    TenantHeader:
      name: X-Tenant-ID
      in: header
      required: true
      schema: { type: string, minLength: 26, maxLength: 26 }
...
  schemas:
...
    RetrieveRequest:
      type: object
      required: [query]
      properties:
        query: { type: string }
        top_k: { type: integer, default: 10 }
        sources: { type: array, items: { type: string } }
        channels: { type: array, items: { type: string } }
        document_ids: { type: array, items: { type: string } }
        privacy_mode: { type: string, enum: [no-ai, internal, public, local-only] }
        skill_id: { type: string }
        device_tier: { type: string, enum: [low, mid, high] }
    RetrieveResponse:
      type: object
      properties:
        hits:
          type: array
          items: { $ref: '#/components/schemas/RetrieveHit' }
        policy: { $ref: '#/components/schemas/RetrievePolicy' }
        timings: { $ref: '#/components/schemas/RetrieveTimings' }
        trace_id: { type: string }
        prefer_local: { type: boolean }
        prefer_local_reason: { type: string }
    RetrieveHit:
      type: object
      properties:
        id: { type: string }
        score: { type: number }
        tenant_id: { type: string }
        source_id: { type: string }
        document_id: { type: string }
        block_id: { type: string }
        title: { type: string }
        uri: { type: string }
        privacy_label: { type: string }
        text: { type: string }
        connector: { type: string }
        metadata: { type: object, additionalProperties: true }
        privacy_strip: { type: object, additionalProperties: true }
    RetrievePolicy:
      type: object
      properties:
        applied: { type: array, items: { type: string } }
        degraded: { type: array, items: { type: string } }
        blocked_count: { type: integer }
        cache_hit: { type: boolean }
        privacy_mode: { type: string }
    RetrieveTimings:
      type: object
      properties:
        vector_ms: { type: integer }
        bm25_ms: { type: integer }
        graph_ms: { type: integer }
        memory_ms: { type: integer }
        merge_ms: { type: integer }
        rerank_ms: { type: integer }
    BatchRequest:
      type: object
      required: [requests]
      properties:
        requests:
          type: array
          maxItems: 32
          items: { $ref: '#/components/schemas/RetrieveRequest' }
        max_parallel: { type: integer, default: 8 }
    BatchResponse:
      type: object
      properties:
        results:
          type: array
          items:
            type: object
            properties:
              index: { type: integer }
              response: { $ref: '#/components/schemas/RetrieveResponse' }
              error: { type: string }
...
    DLQEntry:
      type: object
      properties:
        id: { type: string }
        tenant_id: { type: string }
        topic: { type: string }
        partition: { type: integer }
        offset: { type: integer }
        payload: { type: string }
        reason: { type: string }
        attempts: { type: integer }
        created_at: { type: string, format: date-time }
...
    DLQListResponse:
      type: object
      # Round-11 Task 17.
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/DLQEntry' }
        next_page_token: { type: string }
...
    BatchRetrieveRequest:
      type: object
      required: [requests]
      properties:
        requests:
          type: array
          items: { $ref: '#/components/schemas/RetrieveRequest' }
        diversity: { type: number, format: float }
...
paths:
...
  /v1/retrieve/batch:
    post:
      summary: Run multiple retrieves concurrently in one request.
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
      requestBody:
        required: true
        content:
          application/json:
            # BatchRequest is the historical envelope; BatchRetrieveRequest
            # mirrors internal/retrieval.BatchRetrieveRequest.
            schema:
              oneOf:
                - { $ref: '#/components/schemas/BatchRequest' }
                - { $ref: '#/components/schemas/BatchRetrieveRequest' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/BatchResponse' } } } }
        '400': { description: Bad request }
        '413': { description: Batch too large }
...
  /v1/admin/dlq:
    get:
      summary: List dead-letter messages.
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/DLQListResponse' } } }
  /v1/admin/dlq/{id}:
    get:
      summary: Get a single DLQ entry.
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
  /v1/admin/dlq/{id}/replay:
    post:
      summary: Replay a DLQ entry into the main ingest topic.
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '202': { description: Accepted }
        '409': { description: Max retries exceeded }
...